Made the code Readable and Modularized it
